# perf(html/parser): improve lexer perf