<!-- Note: CI script will automatically rebase your PR so please do not rebase unless required -->

<!-- Thanks for submitting a pull request! Please provide enough information so that others can review your pull request. -->

**Description:**

Improve perf of lexer + simplify code, there are still more places, but I will improve it after merge it.

Parsing stackoverflow from benches:

Before:
```
html/lexer/stackoverflow_com_17_05_2022
                        time:   [71.986 ms 72.358 ms 72.837 ms]
                        change: [-2.5210% -1.8364% -1.0132%] (p = 0.00 < 0.05)
                        Performance has improved.
```

After:
```
html/lexer/stackoverflow_com_17_05_2022
                        time:   [11.193 ms 11.224 ms 11.260 ms]
                        change: [-84.599% -84.488% -84.390%] (p = 0.00 < 0.05)
```

**BREAKING CHANGE:**

No

**Related issue (if exists):**

No